+++ CHANGELOG.md
@@ -8,6 +8,13 @@
 ### Changed
 - Changed login message to include the username you have logged in as, if possible.
 
+### Fixed
+- Various fixes for 'container-registry' commands:
+  - Added missing container-registry columns
+  - The garbage collection days and time are now randomly generated to be a random working day during hours 10:00 - 16:00
+  - Fixed listing container-registry token scopes
+  - Fixed various panics related to nil ExpiryDate on container-registry tokens including #505
+
 ## [v6.8.1] (February 2025)
 
 ### Added